chore: #4 컴포넌트 구조
